- extract score card data into a hook
